move Visibility into data folder
