+++ lib/mastodon_dart.dart
@@ -20,8 +20,8 @@ export 'src/data/push_subscription.dart';
 export 'src/data/relationship.dart';
 export 'src/data/results.dart';
 export 'src/data/scheduled_status.dart';
-export 'src/data/shared/visibility.dart';
 export 'src/data/status.dart';
 export 'src/data/tag.dart';
 export 'src/data/user_follows_list.dart';
+export 'src/data/visibility.dart';
 export 'src/mastodon.dart';
